fix(build): resolve freestanding linking errors and compiler warnings
- Implement freestanding `mul_div_64` via inline `mulq`/`divq` in `hpet.cpp` to eliminate undefined reference to `__udivti3`
- Add explicit `PHDRS` headers in `linker.ld` to prevent RWX segment warnings
- Add `.note.GNU-stack` sections to `boot.asm` and `isr_stubs.asm` to enforce non-executable stack metadata
- Fix signed/unsigned comparison warning in `terminal_app.cpp` history loop

+++ OS/src/apps/terminal_app.cpp
@@ -475,7 +475,7 @@ void TerminalApp::execute_command(const char* cmd) {
         }
     } else if (str_equals(cmd, "history")) {
         print_line("Command History:");
-        for (size_t i = 0; i < m_history_count; ++i) {
+        for (int32_t i = 0; i < m_history_count; ++i) {
             print_char('0' + ((i + 1) / 10) % 10);
             print_char('0' + ((i + 1) % 10));
             print_string("  ");
